requirements.txt
django
django-widget-tweaks
mysqlclient
PyMySQL
cryptography
django-backup
django-select2
openpyxl
reportlab
python-decouple
djangorestframework